feat: check commits instead of "Your branch is up-to-date."

+++ src/Faramond/FaramondManager.php
@@ -54,25 +54,50 @@ class FaramondManager
 
         if($branch == null) {
             $branch = $this->branch;
-
         }
+
+        if(!$this->updatesAvailable($branch, $verbose)) {
+
+            $this->esit[] = [
+                "description" => "Already up to date.",
+                "command" => "",
+                "result" => null
+            ];
+
+            return $this->completeDeploy();
+        }
+
         $this->esit[] = $this->execCommand("Activating manteinance mode","cd $this->root_dir && php artisan down",$verbose);
         $this->esit[] = $this->execCommand("Removing not-in-repo files","cd $this->root_dir && $this->git clean -f",$verbose);
         $this->esit[] = $this->execCommand("Fetch git refs","cd $this->root_dir && $this->git fetch",$verbose);
         $this->esit[] = $this->execCommand("Resetting repo to default state","cd $this->root_dir && $this->git checkout .",$verbose);
         $this->esit[] =  $this->execCommand("Ensuring we are on the correct branch","cd $this->root_dir && $this->git checkout ".$branch,$verbose);
-
-        if(strpos(end($this->esit)['result'], "Your branch is up-to-date") !== false) {
-            return $this->completeDeploy($this->esit, $verbose);
-        }
-
         $this->esit[] = $this->execCommand("Pulling from upstream","cd $this->root_dir && $this->git pull origin ".$branch,$verbose);
         $this->esit[] = $this->execCommand("Creating composer temp directory","cd $this->root_dir && mkdir -p composer_temp",$verbose);
         putenv('COMPOSER_HOME='.$this->root_dir."/composer_temp");
         $this->esit[] = $this->execCommand("Updating composer","cd $this->root_dir && $this->composer update",$verbose);
         $this->esit[] = $this->execCommand("Removing composer temp directory","cd $this->root_dir && rm -r composer_temp",$verbose);
         $this->esit[] = $this->execCommand("Running migrations","cd $this->root_dir && php artisan migrate --force",$verbose);
-        return $this->completeDeploy($this->esit, $verbose);
+        $this->esit[] = $this->execCommand("Deactivating manteinance mode","cd $this->root_dir && php artisan up",$verbose);
+
+        return $this->completeDeploy();
+    }
+
+    /**
+     * Check if the selected branch last commit is the same of the local HEAD
+     * @param $branch
+     * @param bool $verbose
+     * @return int
+     */
+    private function updatesAvailable ($branch, $verbose = false) {
+
+        $this->esit[] = $this->execCommand("Get current commit", "cd $this->root_dir && git rev-parse HEAD", $verbose);
+        $currentCommit = end($this->esit)['result'];
+
+        $this->esit[] = $this->execCommand("Get reference branch last commit.", "$this->git rev-parse origin/".$branch, $verbose);
+        $originCommit = end($this->esit)['result'];
+
+        return strcmp($currentCommit, $originCommit);
     }
 
     /**
@@ -90,7 +115,7 @@ class FaramondManager
         return [
             "description" => $description,
             "command" => $command,
-            "result" => $result
+            "result" => trim($result)
         ];
     }
 
@@ -99,8 +124,7 @@ class FaramondManager
      * @param array $esit
      * @return array
      */
-    private function completeDeploy(array $esit, $verbose = false) {
-        $esit[] = $this->execCommand("Deactivating manteinance mode","cd $this->root_dir && php artisan up",$verbose);
-        return $esit;
+    private function completeDeploy() {
+        return $this->esit;
     }
 }
